+++ .github/actions/run-tests/test_runner/argument_parser.py
@@ -6,6 +6,7 @@ def parseArguments():
 	parser = argparse.ArgumentParser()
 
 	parser.add_argument('-v', '--verbose', action='store_true', help='Do not hide the output of intermediate commands')
+	parser.add_argument('--debug-python', action='store_true', help='Activate the python debugger to develop the test environments.')
 
 	group = parser.add_argument_group('Docker image management', 'Manage the docker images involved in the test routines.')