+++ CHANGELOG.md
@@ -35,6 +35,8 @@
   [#1710](https://github.com/nextcloud/cookbook/pull/1710) @dependabot @christianlupus
 - Update Psalm and fix some introduced issues
   [#1707](https://github.com/nextcloud/cookbook/pull/1707) @christianlupus
+- Allow remote debugging of the test python scripts
+  [#1713](https://github.com/nextcloud/cookbook/pull/1713) @christianlupus
 
 
 ## 0.10.2 - 2023-03-24